✨(frontend) update sender of draft message
Previously it wasn't possible to change the sender of a draft message.
Now we allow that and in this case we remove the current draft and create
a new one in the new mailbox. After that we redirect the user to the view of
the new draft to continue it's edit!