♻️(backend) make generic user list endpoint
Currently the endpoint to search users is nested under maildomain. So since we
add the new endpoint to manage maildomain accesses, we need more granularity
on this endpoint and its current implementation does not allow that. So we move
it into the UserViewset. This endpoint is accessible to superuser and maildomain
admins only. Superusers are allowed to search into the whole user database to
assign any user as maildomain admin while maildomain admin are only allowed to
search users with maildomain or mailbox access in their scope.