🚸(frontend) display percentage as import progress
Currently we are displaying the number of message imports over the total to
import. According to the import source, the total count cannot be exact as some
message can be count twice, etc. So in order to reduce confusion about the
final import state we now display progress as percentage.